Added code to prepare the dataset:
Create an array of indices, shuffle it and divide it in train (80%) and test (20%), saving to file the configuration.

Divide the datasets based on the run using the indices.

Then create two tensors for both train and test set, one with the list of prox_values as input [size: (n_robot * n_step * n_run) x n_sensors (= 7)] and the other with the speed as output [size: (n_robot * n_step * n_run) x speed (= 1)].