Fixed ManualController:
fixed previous error that occurred in the case of all-sensors net input. Now the controller, instead of using the intensities of the sensors to decide if go backwards or forward, uses distances obtained through interpolation using a function that maps the perceived sensing to distances. All the data used in the mapping function are obtained in the experiment performed in sensor_to_distances.py.